refactor tensor_parallel and comm_utils
