+++ vllm_ascend/ops/moe_dispatcher/token_dispatcher.py
@@ -25,18 +25,17 @@ from abc import ABC, abstractmethod
 from typing import Any, Dict, Optional
 
 import torch
+import torch.distributed
+import torch.distributed as dist
 import torch_npu
 from vllm.distributed.parallel_state import get_ep_group
 
 import vllm_ascend.envs as envs_ascend
 from vllm_ascend.distributed.parallel_state import get_mc2_group
-from vllm_ascend.distributed.tensor_parallel import \
-    gather_from_sequence_parallel_region
-from vllm_ascend.ops.comm_utils import async_all_to_all
 from vllm_ascend.utils import AscendSocVersion, get_ascend_soc_version
 
 _Dispatchers: Dict[str, Any] = {}
-
+COMM_STREAM = None
 
 def _register_token_dispatcher(dispatcher: Any):
     _Dispatchers[dispatcher.__class__.__name__] = dispatcher
@@ -61,6 +60,91 @@ def setup_token_dispatchers(ep_size: int, **kwargs):
             _register_token_dispatcher(TokenDispatcherWithAll2AllV(**kwargs))
         if "TokenDispatcherWithMC2" not in existing_dispatchers:
             _register_token_dispatcher(TokenDispatcherWithMC2(**kwargs))
+
+
+def gather_from_sequence_parallel_region(
+    input_,
+    group,
+    output_split_sizes=None,
+):
+    """Wrapper for autograd function: forward: AG, backward: RS <first dim>
+
+        Args:
+            input_tensor (torch.Tensor):
+                A tensor to be gathered.
+            output_split_sizes (List[int], optional):
+                A list specifying the sizes of the output splits along the first dimension.
+                If None, equal splitting is assumed. Default: None.
+
+        Returns:
+            torch.Tensor: Gathered tensor.
+        """
+    world_size = torch.distributed.get_world_size(group)
+    # Bypass the function if we are using only 1 GPU.
+    if world_size == 1:
+        return input_
+
+    dim_size = list(input_.size())
+    if output_split_sizes is None:
+        dim_size[0] = dim_size[0] * world_size
+
+        output = torch.empty(dim_size,
+                             dtype=input_.dtype,
+                             device=torch.npu.current_device())
+        torch.distributed.all_gather_into_tensor(output,
+                                                 input_.contiguous(),
+                                                 group=group)
+    else:
+        dim_size[0] = sum(output_split_sizes)
+        output = torch.empty(dim_size,
+                             dtype=input_.dtype,
+                             device=torch.npu.current_device())
+        output_tensor_list = list(
+            torch.split(output, output_split_sizes, dim=0))
+        torch.distributed.all_gather(output_tensor_list, input_, group=group)
+
+    return output
+
+
+def async_all_to_all(input_,
+                     output_split_sizes,
+                     input_split_sizes,
+                     group,
+                     event=None):
+    if output_split_sizes is None:
+        # Equal split (all2all)
+        a2a_out = torch.empty_like(input_)
+    else:
+        # Unequal split (all2all-v)
+        a2a_out = input_.new_empty(
+            size=[sum(output_split_sizes)] + list(input_.size()[1:]),
+            dtype=input_.dtype,
+            device=torch.npu.current_device(),
+        )
+
+    if event:
+        # multi stream wait event
+        global COMM_STREAM
+        if COMM_STREAM is None:
+            COMM_STREAM = torch_npu.npu.Stream(
+                device=torch.npu.current_device())
+        with torch_npu.npu.stream(COMM_STREAM):
+            event.wait()
+            handle = dist.all_to_all_single(
+                a2a_out,
+                input_.contiguous(),
+                output_split_sizes=output_split_sizes,
+                input_split_sizes=input_split_sizes,
+                group=group,
+                async_op=True)
+    else:
+        handle = dist.all_to_all_single(a2a_out,
+                                        input_.contiguous(),
+                                        output_split_sizes=output_split_sizes,
+                                        input_split_sizes=input_split_sizes,
+                                        group=group,
+                                        async_op=True)
+    return input_, a2a_out, handle
 
 
 class MoETokenDispatcher(ABC):
